Update While.java
while - condition checked first. 
the while loop is java's most fundamental loop statement. It repeats a statement or block while its controlling expression is TRUE. th econdition can be any boolean expression. the body of the loop will be executed as long as the conditional expression is true. When condition becomes false, control passes to the next line of code ..