Update DoWhile.java
Sometimes it is desirable to execute the body of a loop at least once, even if the condition expression is false to begin with. in other words, there are times when we would like to test the termination expression at the end of the loop rather than at the beginning. Fortunatly, java suplies a loop that does just that: the Do-while loop always its body at least once, because its conditional expression is at the bottom of the loop.
do{
//body of loop
} while(condition);

*runs once, then checks condition.